+++ src/main/java/racingcar/carAction/CarStop.java
@@ -1,6 +1,6 @@
 package racingcar.carAction;
 
-import racingcar.domian.car.Car;
+import racingcar.domain.car.Car;
 
 public class CarStop implements CarAction {
