+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,4 +1,4 @@
-package racingcar.domian.car;
+package racingcar.domain.car;
 
 public class Car implements Comparable<Car> {
     private final Name carName;
